Update and rename matrix_multiplication.py to test_matrix_multiplication.py
